Updated README
To take a look at my portfolio via browser, all you need in accessing it is to click on README in the GitHub page.

To view my code for correction, comments and/or compare it with a given project of the same or so, you can download my files and open them in whatever codding application of your choice such as VS code, Atom, Text Editor, or directly in this GitHub page by clicking the CSS folder or the html folder. This will give you access to the main code files.

To make things clear, the CSS folder contains 2 files. One is the styles.css file where all designs are made for all my html pages or website. The second file you will find is the style.production.css which was generated from the styles.css code after preprocessing.

In case you have any questions or comment don’t  hessite to contact me through my email via the contact page or even here at GitHub thanks.